6차 세미나 과제 완료 #5
- 경고 문구 띄우기, 조건부 버튼 활성화 구현은 완료함.
- databinding 안 썼음
- RegisterActivity에서 통신에 성공한 경우는 문제 없지만 실패 시 NPE가 뜨고 앱이 죽어버림
- LoginActivity에서 통신 성공 시 ToastMessage를 띄우는 코드에서 에러가 남. NPE가 뜨고 앱이 죽어버림.
