--- app/src/main/java/com/example/sopt_seminar_2/HomeDto.kt
@@ -1,5 +1,0 @@
-package com.example.sopt_seminar_2
-
-data class HomeDto(
-    val items: List<HomeModel>
-)
