--- app/src/main/java/com/example/sopt_seminar_2/HomeModel.kt
@@ -1,8 +1,0 @@
-package com.example.sopt_seminar_2
-
-data class HomeModel(
-    val id: String,
-    val position: String,
-    val name: String,
-    val imgUrl: String,
-)
