+++ app/src/main/java/com/seminar_6/data/api/ApiGetUserList.kt
@@ -1,0 +1,18 @@
+package com.seminar_6.data.api
+
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.json.Json
+import okhttp3.MediaType.Companion.toMediaType
+import retrofit2.Retrofit
+
+object ApiGetUserList {
+
+    val retrofit: Retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl("https://reqres.in/")
+            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
+            .build()
+    }
+
+    inline fun <reified T> create(): T = retrofit.create<T>(T::class.java)
+}
