+++ app/src/main/java/com/seminar_6/data/api/ApiLoginRegister.kt
@@ -1,0 +1,22 @@
+package com.seminar_6.data.api
+
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.json.Json
+import okhttp3.MediaType.Companion.toMediaType
+import retrofit2.Retrofit
+
+object ApiLoginRegister {
+
+    val retrofit: Retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl("http://3.39.169.52:3000/")
+            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
+            .build()
+    }
+
+    inline fun <reified T> create(): T = retrofit.create<T>(T::class.java)
+
+    object ServicePool {
+        val loginRegisterService = create<LoginRegisterService>()
+    }
+}
