+++ app/src/main/java/com/seminar_6/data/api/UserListService.kt
@@ -1,0 +1,13 @@
+package com.seminar_6.data.api
+
+import com.seminar_6.data.model.UserListDto.ResponseUserListDto
+import retrofit2.Response
+import retrofit2.http.GET
+
+interface UserListService {
+    @GET("api/users?page=2")
+    suspend fun getUserList(
+    ): Response<ResponseUserListDto>
+}
+
+
