+++ app/src/main/java/com/seminar_6/data/model/LoginDto/ResponseLoginDto.kt
@@ -1,0 +1,31 @@
+package com.seminar_6.data.model.LoginDto
+
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseLoginDto(
+    @SerialName("message")
+    val message: String,
+    @SerialName("result")
+    val result: Result,
+    @SerialName("status")
+    val status: Int
+) {
+    @Serializable
+    data class Result(
+        @SerialName("bio")
+        val bio: String?,
+        @SerialName("email")
+        val email: String,
+        @SerialName("id")
+        val id: Int,
+        @SerialName("name")
+        val name: String,
+        @SerialName("password")
+        val password: String,
+        @SerialName("profileImage")
+        val profileImage: String?
+    )
+}
