+++ app/src/main/java/com/seminar_6/presentation/UserListAdapter.kt
@@ -1,0 +1,75 @@
+package com.sopt_seminar_4.View
+
+import android.content.Context
+import android.util.TypedValue
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.CenterCrop
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.seminar_6.data.model.UserListDto.ResponseUserListDto
+import com.seminar_6.databinding.RecyclerviewElementBinding
+
+class UserListAdapter :
+    ListAdapter<ResponseUserListDto.Data, UserListAdapter.ItemViewHolder>(differ()) {
+
+
+    inner class ItemViewHolder(val binding: RecyclerviewElementBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+
+        fun bind(userList: ResponseUserListDto.Data) {
+
+            binding.tvEmail.text = userList.email
+            binding.tvUserName.text = userList.firstName
+
+            Glide
+                .with(binding.thumbnailImage.context)
+                .load(userList.avatar)
+                .transform(CenterCrop(), RoundedCorners(dpToPx(binding.thumbnailImage.context, 12)))
+                .into(binding.thumbnailImage)
+
+
+        }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemViewHolder {
+        val inflater = LayoutInflater.from(parent.context)
+        val binding = RecyclerviewElementBinding.inflate(inflater)
+
+        return ItemViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: ItemViewHolder, position: Int) {
+        holder.bind(currentList[position])
+    }
+
+    private fun dpToPx(context: Context, dp: Int): Int {
+        return TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_DIP,
+            dp.toFloat(),
+            context.resources.displayMetrics
+        ).toInt()
+    }
+
+
+    class differ() : DiffUtil.ItemCallback<ResponseUserListDto.Data>() {
+        override fun areItemsTheSame(
+            oldItem: ResponseUserListDto.Data,
+            newItem: ResponseUserListDto.Data
+        ): Boolean {
+            return oldItem.id == newItem.id
+        }
+
+        override fun areContentsTheSame(
+            oldItem: ResponseUserListDto.Data,
+            newItem: ResponseUserListDto.Data
+        ): Boolean {
+            return oldItem == newItem
+        }
+
+    }
+}
