--- app/src/main/java/com/sopt_seminar_4/Data/API/ApiGetUserList.kt
@@ -1,18 +1,0 @@
-package com.sopt_seminar_4.Data.API
-
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import retrofit2.Retrofit
-
-object ApiGetUserList {
-
-    val retrofit: Retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl("https://reqres.in/")
-            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
-            .build()
-    }
-
-    inline fun <reified T> create(): T = retrofit.create<T>(T::class.java)
-}
