--- app/src/main/java/com/sopt_seminar_4/Data/DTO/RegisterDTO/RequestRegisterDTO.kt
@@ -1,15 +1,0 @@
-package com.sopt_seminar_4.Data.DTO.RegisterDTO
-
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestRegisterDTO(
-    @SerialName("email")
-    val email: String,
-    @SerialName("name")
-    val name: String,
-    @SerialName("password")
-    val password: String
-)
