+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
 <resources>
-    <string name="app_name">sopt_seminar_4</string>
+    <string name="app_name">seminar_6</string>
+    <string name="not_match_password_condition">비밀번호 조건에 맞지 않습니다</string>
+    <string name="enter_your_password">비밀번호를 입력해주세요</string>
+    <string name="password">비밀번호</string>
+    <string name="enter_your_email">이메일을 입력해주세요</string>
+    <string name="email">이메일</string>
 </resources>
